Incremental improvement in shopping cart DAO.
SVN-Revision: 13326

+++ src/org/rti/webcgh/service/dao/hibernate/unit_test/HibernateShoppingCartDaoTester.java
@@ -51,6 +51,7 @@ ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 package org.rti.webcgh.service.dao.hibernate.unit_test;
 
 
+import org.rti.webcgh.domain.Plot;
 import org.rti.webcgh.domain.ShoppingCart;
 import org.rti.webcgh.service.dao.hibernate.HibernateShoppingCartDao;
 import org.springframework.context.ApplicationContext;
@@ -68,15 +69,56 @@ public final class HibernateShoppingCartDaoTester extends TestCase {
      * Test all methods.
      */
     public void testAllMethods() {
+    	
+    	// Setup for tests
     	ApplicationContext ctx = new ClassPathXmlApplicationContext(
     		"org/rti/webcgh/service/dao/hibernate/unit_test/beans.xml");
     	HibernateShoppingCartDao dao = (HibernateShoppingCartDao)
     		ctx.getBean("shoppingCartDao");
     	String user = "user";
+    	
+    	// Create a shopping cart and add 2 plots
     	ShoppingCart c1 = new ShoppingCart(user);
+    	c1.add(new Plot("plot1"));
+    	c1.add(new Plot("plot2"));
     	dao.save(c1);
+    	
+    	// Test 1
     	ShoppingCart c2 = dao.load(user);
     	assertNotNull(c2);
+    	assertEquals(c1.getUserName(), c2.getUserName());
+    	assertEquals(2, c2.getPlots().size());
+    	
+    	// Test 2 - Add new plot and update
+    	Plot plot3 = new Plot("plot3");
+    	c1.add(plot3);
+    	dao.update(c1);
+    	c2 = dao.load(user);
+    	assertEquals(3, c2.getPlots().size());
+    	
+    	// Test 3 - Change name of a plot
+    	String newName = "new name";
+    	plot3.setName(newName);
+    	dao.update(c1);
+    	c2 = dao.load(user);
+    	boolean found = false;
+    	for (Plot p : c2.getPlots()) {
+    		if (newName.equals(p.getName())) {
+    			found = true;
+    			break;
+    		}
+    	}
+    	if (!found) {
+    		fail();
+    	}
+    	
+    	// Test 4 - Remove a plot and update
+    	c1.remove(plot3);
+    	dao.update(c1);
+    	c2 = dao.load(user);
+    	assertEquals(2, c2.getPlots().size());
+    	
+    	// Test 5 - Delete shopping cart
     	dao.delete(c1);
     	c2 = dao.load(user);
     	assertNull(c2);
